Add "Do Not Kill" whitelist support for apps [2/2]
This provides a UI for the lowmemorykiller whitelist module.
Requires compatible kernel.

Check "Do not kill" under installed app details to whitelist the
app. Maximum of 12 apps can be added to the whitelist.

Whitelisted apps appear on an additional tab.

Whitelisting of platform signed apps is prohibited.

Interface is disabled if the kernel module is not present.

Change-Id: I0e4bc16719c01396897b624adf2bfbcd9fccd3a9